Using jscodeshift
