+++ package.json
@@ -8,9 +8,7 @@
   "license": "SEE LICENSE IN https://tc39.github.io/ecma262/#sec-copyright-and-software-license",
   "homepage": "https://iddan.github.io/proposal-function-expression-decorators/",
   "dependencies": {
-    "acorn": "^5.5.0",
-    "ecmarkup": "^3.12.0",
-    "recast": "^0.14.4"
+    "ecmarkup": "^3.12.0"
   },
   "devDependencies": {
     "@alrra/travis-scripts": "^3.0.0"
